backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pymongo[srv]
motor
langchain-text-splitters
pymupdf
python-multipart
python-dotenv
supabase
passlib
bcrypt==4.0.1
pyjwt
